+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,10 +5,11 @@
 ## 1.17.0 - (March 29, 2023)
 
 * Changed
-  * Added test for `terra-slide-panel`.
   * Updated docs for `terra-slide-group`.
 * Added
   * Added package-lock.json
+  * Added test for `terra-slide-panel`.
+  * Added test for `terra-slide-panel-manager`.
 
 ## 1.16.0 - (March 1, 2023)